Image_steganography/requirements.txt
Pillow==9.5.0
pycryptodome==3.17
